ts-cli: move ndjson and config tests next to source files
Committed-By-Agent: claude
